CRUD usuarios productos categorias

+++ Controllers/ProductoController.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using TiendaVirtualYanten.Data;
 using TiendaVirtualYanten.Models;
@@ -13,51 +13,61 @@ namespace TiendaVirtualYanten.Controllers
         {
             _context = context;
         }
+
         public IActionResult Index()
         {
             var productos = _context.Productos
                 .Include(p => p.Categoria)
                 .ToList();
-
             return View(productos);
         }
+
         public IActionResult Create()
         {
+            ViewBag.Categorias = _context.Categorias.ToList();
             return View();
         }
 
         [HttpPost]
-
         public IActionResult Create(Producto producto)
         {
-            _context.Productos.Add(producto);
-            _context.SaveChanges();
-
-            return RedirectToAction("Index");
+            if (ModelState.IsValid)
+            {
+                _context.Productos.Add(producto);
+                _context.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            ViewBag.Categorias = _context.Categorias.ToList();
+            return View(producto);
         }
+
         public IActionResult Edit(int id)
         {
             var producto = _context.Productos.Find(id);
+            if (producto == null) return RedirectToAction("Index");
             ViewBag.Categorias = _context.Categorias.ToList();
-
             return View(producto);
         }
 
         [HttpPost]
         public IActionResult Edit(Producto producto)
         {
-            _context.Productos.Update(producto);
-            _context.SaveChanges();
-
-            return RedirectToAction("Index");
+            if (ModelState.IsValid)
+            {
+                _context.Productos.Update(producto);
+                _context.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            ViewBag.Categorias = _context.Categorias.ToList();
+            return View(producto);
         }
+
         public IActionResult Delete(int id)
         {
             var producto = _context.Productos.Find(id);
-
+            if (producto == null) return RedirectToAction("Index");
             _context.Productos.Remove(producto);
             _context.SaveChanges();
-
             return RedirectToAction("Index");
         }
     }
